arcade-ai/requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
pydantic==1.10.13
python-dotenv==1.0.0
arcadepy==1.4.0
httpx==0.23.0
requests==2.31.0